+++ lib/domain/usecase/play_list_usecase.dart
@@ -1,6 +1,6 @@
 import 'package:oz_player/data/dto/play_list_dto.dart';
 import 'package:oz_player/domain/entitiy/play_list_entity.dart';
-import 'package:oz_player/domain/repository/play_list_repository.dart';
+import 'package:oz_player/domain/repository/saved/play_list_repository.dart';
 
 class PlayListUsecase {
   final PlayListRepository _repository;
